requirements.txt
aiohttp==3.9.3
fastapi==0.110.0
python-dotenv==1.0.1
Requests==2.31.0
pydantic==2.6.4
uvicorn==0.29.0
googlesearch-python
httpx
